Fixed #116: Animation Jittering when transferring from faster conveyors to slower ones, completed a pending TODO (insertion check), some method cleanup

+++ common/net/teamio/taam/content/conveyors/TileEntityConveyorHopper.java
@@ -353,7 +353,7 @@ public class TileEntityConveyorHopper extends BaseTileEntity implements IConveyo
 	}
 
 	@Override
-	public int getSpeedsteps() {
+	public byte getSpeedsteps() {
 		return 1;
 	}
