+++ common/net/teamio/taam/conveyors/api/IConveyorAwareTE.java
@@ -8,7 +8,7 @@ public interface IConveyorAwareTE {
 	boolean canSlotMove(int slot);
 	boolean isSlotAvailable(int slot);
 	int getMovementProgress(int slot);
-	int getSpeedsteps();
+	byte getSpeedsteps();
 	
 	IItemFilter getSlotFilter(int slot);
